+++ src/error.rs
@@ -10,14 +10,14 @@ use std::io;
 #[derive(Debug)]
 /// A custom Goblin error
 pub enum Error {
-    /// An IO based error
-    IO(io::Error),
-    /// An error emanating from reading and interpreting bytes
-    Scroll(scroll::Error),
     /// The binary is malformed somehow
     Malformed(String),
     /// The binary's magic is unknown or bad
     BadMagic(u64),
+    /// An error emanating from reading and interpreting bytes
+    Scroll(scroll::Error),
+    /// An IO based error
+    IO(io::Error),
 }
 
 impl error::Error for Error {